Calculatrice V1.2

+++ Calculatrice/app/src/main/java/fr/doranco/calculatrice/Calculateur.java
@@ -47,7 +47,7 @@ public class Calculateur {
         }
     }
 
-    public float getResultat() {
+    public int getResultat() {
         return resultat;
     }
